+++ cmd/clogin/main.go
@@ -25,7 +25,7 @@ import (
 	_ "gorancid/pkg/parse/nxos"
 )
 
-const version = "0.4.0"
+const version = "0.4.1"
 
 func main() {
 	var (