+++ cmd/control-rancid/main.go
@@ -19,7 +19,7 @@ import (
 	"gorancid/pkg/parse"
 )
 
-const version = "0.4.0"
+const version = "0.4.1"
 
 func main() {
 	var (