+++ cmd/rancid/main.go
@@ -15,7 +15,7 @@ import (
 	"gorancid/pkg/parse"
 )
 
-const version = "0.4.0"
+const version = "0.4.1"
 
 func main() {
 	var (